requirements.txt
aiogram
tortoise-orm[psycopg]
python-decouple